moved function for concepticon definition to add_concepticon
